+++ models/datasources.py
@@ -2,7 +2,8 @@
 
 import os
 from PyQt4.QtCore import Qt, pyqtSignal
-from ..datasource.netcdf_groundwater import NetcdfGroundwaterDataSource
+from ..datasource.netcdf_groundwater_h5py import \
+    NetcdfGroundwaterDataSourceH5py
 from ..datasource.netcdf import NetcdfDataSource
 from .base import BaseModel
 from .base_fields import CheckboxField, ValueField
@@ -67,7 +68,7 @@ class DataSourceLayerManager(object):
     """
     type_ds_mapping = {
         'netcdf': NetcdfDataSource,
-        'netcdf-groundwater': NetcdfGroundwaterDataSource,
+        'netcdf-groundwater': NetcdfGroundwaterDataSourceH5py,
     }
 
     def __init__(self, ds_type, file_path):